backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
torch
torchvision
opencv-python
facenet-pytorch
reportlab
numpy
pillow
